added telemetry, fix db init script

+++ scripts/init_db.sh
@@ -22,28 +22,29 @@ DB_NAME=${POSTGRES_DB:=newsletter}
 DB_PORT=${POSTGRES_PORT:=5432}
 DB_HOST=${POSTGRES_HOST:=localhost}
 
-if [ ! "$(docker ps -a -q -f name=zero2prod-postgresql)" ]; then
-  if [ "$(docker ps -a -q -f status=exited -f name=zero2prod-postgresql)" ]; then
-    docker rm zero2prod-postgresql
-  fi
+if [[ -z "${SKIP_DOCKER}" ]]; then
+  if [ ! "$(docker ps -a -q -f name=zero2prod-postgresql)" ]; then
+    if [ "$(docker ps -a -q -f status=exited -f name=zero2prod-postgresql)" ]; then
+      docker rm zero2prod-postgresql
+    fi
 
-  docker run \
-    -e POSTGRES_USER="${DB_USER}" \
-    -e POSTGRES_PASSWORD="${DB_PASSWORD}" \
-    -e POSTGRES_DB="${DB_NAME}" \
-    -p "${DB_PORT}":5432 \
-    --name zero2prod-postgresql \
-    -d postgres \
-    postgres -N 1000
+    docker run \
+      -e POSTGRES_USER="${DB_USER}" \
+      -e POSTGRES_PASSWORD="${DB_PASSWORD}" \
+      -e POSTGRES_DB="${DB_NAME}" \
+      -p "${DB_PORT}":5432 \
+      --name zero2prod-postgresql \
+      -d postgres \
+      postgres -N 1000
+  fi
 fi
 
 export PGPASSWORD="${DB_PASSWORD}"
 until psql -h "${DB_HOST}" -U "${DB_USER}" -p "${DB_PORT}" -d "postgres" -c '\q'; do
-  >&2 echo "Postgress is still unavailable - sleeping for 1 second"
+  echo >&2 "Postgress is still unavailable - sleeping for 1 second"
   sleep 1
 done
 
 export DATABASE_URL="postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}"
 sqlx database create
 sqlx migrate run
-
